+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### next
+- sort, accessible through the menu - Fix #31
+
 <a name="v1.2.3"></a>
 ### v1.2.3 - 2023-04-06
 Fix a crash on some combinations of scroll+search